Low-rank data, more details

Assume that matrices $A_i$ in $\sum_{i=1}^{{n}} y_i A_i + S = C,\ \ S\succeq 0,$ are obtained by

\[ A_i = B_i B_i^\top\]

with data matrices $B_i\in{\mathbb R}^{m\times k}$ and with $k \ll n$.

The complexity of $H$ assembly is then reduced from $nm^3$ to $knm^2$.

In particular: If we know that $A_i$ have rank one, the decomposition $A_i = b_i b_i^\top$ is performed by Loraine automatically (datarank = -1).

The model is represented internally via the following struct:

Loraine.Solvers.MyModel — Type

MyModel

Model representing the problem:

\[\begin{aligned} \max {} & b^\top y - b_\text{const} \\ & \sum_{j=1}^n y_j A_{i,j} \preceq C_i \qquad \forall i \in \{1,\ldots,\text{nlmi}\} \\ & C_\text{lin} y \le d_\text{lin} \end{aligned}\]

The fields of the struct as related to the arrays of the above formulation as follows:

The $i$th PSD constraint is of size msize[i] × msisze[i]

The matrix $C_i$ is given by C[i] which should be equal to -A[i,1].

The matrix $A_{i,j}$ is given by -A[i,j+1] as well as myA[(i-1)*n + j].

The vectorization vec(A[i,j+1]) is also given by -AA[i][:,j]

If datarank == -1, $A_{i,j}$ is also equal to -B[i][j,:] * B[i][j,:]'.

The matrix $A_{i,j}$ has nzA[j,i] nonzero entries

The index j = sigmaA[k,i] is the kth matrix $A_{i,j}$ of the largest number of nonzeros.

The first qA[1,i] = qA[2,i] matrices are considered as dense in the computation.